Some cleaning before Liz's ultimate schemes
