--- src/main/java/frc/robot/Commands/AutoCommands/SpeakerShoot.java
@@ -1,32 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.Commands.AutoCommands;
-
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
-import frc.robot.Commands.TeleopCommands.SpeakerScore.PositionToShoot;
-import frc.robot.Commands.TeleopCommands.SpeakerScore.Shoot;
-import frc.robot.Subsystems.arm.Arm;
-import frc.robot.Subsystems.feeder.Feeder;
-import frc.robot.Subsystems.shooter.Shooter;
-import frc.robot.Subsystems.shooter.ShooterConstants;
-import frc.robot.Subsystems.wrist.Wrist;
-import frc.robot.Subsystems.wrist.WristConstants;
-
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class SpeakerShoot extends SequentialCommandGroup {
-  /** Creates a new SpeakerShoot. */
-  public SpeakerShoot(Feeder feeder, Shooter shooter, Wrist wrist, Arm arm) {
-    // Add your commands in the addCommands() call, e.g.
-    // addCommands(new FooCommand(), new BarCommand());
-    addCommands(
-        new PositionToShoot(
-            feeder, shooter, wrist, WristConstants.SUBWOOFER_RAD, ShooterConstants.CLOSE_RPM),
-        new WaitUntilCommand(() -> wrist.atGoal()),
-        new Shoot(feeder, arm, shooter));
-  }
-}
