+++ src/main/java/frc/robot/Commands/AutonomousCommands/DeadReckons/First3Pieces/OnePieceAuto.java
@@ -2,49 +2,43 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Commands.AutonomousCommands.First3Pieces;
+package frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces;
 
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.Subsystems.drive.Drive;
+import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
 import frc.robot.Subsystems.feeder.Feeder;
-import frc.robot.Subsystems.gyro.Gyro;
+import frc.robot.Subsystems.feeder.FeederConstants;
 import frc.robot.Subsystems.shooter.Shooter;
+import frc.robot.Subsystems.shooter.ShooterConstants;
 import frc.robot.Subsystems.wrist.Wrist;
+import frc.robot.Subsystems.wrist.WristConstants;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class OnePieceLeaveAmpSide extends SequentialCommandGroup {
-  /** Creates a new OnePieceLeaveAmpSide. */
-  public OnePieceLeaveAmpSide(
-      Wrist wrist,
-      Feeder feeder,
-      Shooter shooter,
-      Drive drive,
-      double seconds,
-      double speed,
-      Gyro gyro) {
+public class OnePieceAuto extends SequentialCommandGroup {
+  /** Creates a new OnePieceAuto. */
+  public OnePieceAuto(Wrist wrist, Feeder feeder, Shooter shooter) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
+        new WaitCommand(SmartDashboard.getNumber("delay", 0)),
         Commands.runOnce(
             () -> {
-              gyro.zeroYaw();
+              wrist.setGoal(WristConstants.SUBWOOFER_RAD);
+              shooter.setSetpoint(ShooterConstants.CLOSE_RPM);
             },
-            gyro),
-        new OnePieceAuto(wrist, feeder, shooter),
+            shooter,
+            wrist),
+        new WaitUntilCommand(() -> wrist.atGoal()),
+        new WaitUntilCommand(() -> shooter.topAtSetpoint()),
         Commands.runOnce(
             () -> {
-              drive.setRawWithAdjustedHeading(0, speed, 0, gyro.adjustedYaw(60));
+              feeder.setSetpoint(FeederConstants.SPEAKER_RPM);
             },
-            drive),
-        new WaitCommand(seconds),
-        Commands.runOnce(
-            () -> {
-              drive.setRaw(0, 0, 0);
-            },
-            drive));
+            feeder));
   }
 }
