+++ src/main/java/frc/robot/Commands/AutonomousCommands/DeadReckons/First3Pieces/TwoPieceAuto.java
@@ -2,17 +2,17 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Commands.AutonomousCommands.MiddleFieldPieces;
+package frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces;
 
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.OnePieceAuto;
 import frc.robot.Commands.TeleopCommands.Intakes.AllIntakesRun;
 import frc.robot.Commands.TeleopCommands.SpeakerScore.PositionToShoot;
 import frc.robot.Subsystems.actuator.Actuator;
 import frc.robot.Subsystems.drive.Drive;
+import frc.robot.Subsystems.drive.DriveConstants;
 import frc.robot.Subsystems.feeder.Feeder;
 import frc.robot.Subsystems.gyro.Gyro;
 import frc.robot.Subsystems.otbIntake.OTBIntake;
@@ -24,9 +24,9 @@ import frc.robot.Subsystems.wrist.*;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class TwoMidfieldPieceAuto extends SequentialCommandGroup {
+public class TwoPieceAuto extends SequentialCommandGroup {
   /** Creates a new TwoPieceAuto. */
-  public TwoMidfieldPieceAuto(
+  public TwoPieceAuto(
       Drive drive,
       Gyro gyro,
       Wrist wrist,
@@ -36,8 +36,7 @@ public class TwoMidfieldPieceAuto extends SequentialCommandGroup {
       OTBIntake otbIntake,
       UTBIntake utbIntake,
       double seconds,
-      double speedX,
-      double speedY) {
+      double speed) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
@@ -50,14 +49,15 @@ public class TwoMidfieldPieceAuto extends SequentialCommandGroup {
         new ParallelCommandGroup(
             Commands.runOnce(
                 () -> {
-                  drive.setRaw(0, speedY, 0);
+                  wrist.setGoal(0);
                 },
-                drive),
-            Commands.runOnce(
-                () -> {
-                  drive.setRaw(speedX, 0, 0);
-                }),
+                wrist),
             new AllIntakesRun(actuator, otbIntake, utbIntake, feeder, false)),
+        Commands.runOnce(
+            () -> {
+              drive.setRaw(0, speed / DriveConstants.MAX_LINEAR_SPEED_M_PER_SEC, 0);
+            },
+            drive),
         new WaitCommand(seconds),
         Commands.runOnce(
             () -> {
@@ -65,15 +65,6 @@ public class TwoMidfieldPieceAuto extends SequentialCommandGroup {
             },
             drive),
         new AllIntakesRun(actuator, otbIntake, utbIntake, feeder, true),
-        Commands.runOnce(
-            () -> {
-              drive.setRaw(-speedX, 0, 0);
-            },
-            drive),
-        Commands.runOnce(
-            () -> {
-              drive.setRaw(0, -speedY, 0);
-            }),
         new PositionToShoot(
             feeder, shooter, wrist, WristConstants.PODIUM_RAD, ShooterConstants.CLOSE_RPM));
   }
