+++ src/main/java/frc/robot/Commands/AutonomousCommands/DeadReckons/First3Pieces/TwoPieceReturnSub.java
@@ -2,34 +2,31 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Commands.AutonomousCommands.First3Pieces;
+package frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces;
 
 import edu.wpi.first.wpilibj2.command.Commands;
-import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.Commands.TeleopCommands.Intakes.AllIntakesRun;
-import frc.robot.Commands.TeleopCommands.SpeakerScore.PositionToShoot;
 import frc.robot.Subsystems.actuator.Actuator;
+import frc.robot.Subsystems.arm.Arm;
 import frc.robot.Subsystems.drive.Drive;
-import frc.robot.Subsystems.drive.DriveConstants;
 import frc.robot.Subsystems.feeder.Feeder;
 import frc.robot.Subsystems.gyro.Gyro;
 import frc.robot.Subsystems.otbIntake.OTBIntake;
 import frc.robot.Subsystems.shooter.Shooter;
-import frc.robot.Subsystems.shooter.ShooterConstants;
 import frc.robot.Subsystems.utbintake.UTBIntake;
-import frc.robot.Subsystems.wrist.*;
+import frc.robot.Subsystems.wrist.Wrist;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class TwoPieceAuto extends SequentialCommandGroup {
-  /** Creates a new TwoPieceAuto. */
-  public TwoPieceAuto(
+public class TwoPieceReturnSub extends SequentialCommandGroup {
+  /** Creates a new TwoPieceReturnSub. */
+  public TwoPieceReturnSub(
       Drive drive,
       Gyro gyro,
       Wrist wrist,
+      Arm arm,
       Feeder feeder,
       Shooter shooter,
       Actuator actuator,
@@ -42,30 +39,47 @@ public class TwoPieceAuto extends SequentialCommandGroup {
     addCommands(
         Commands.runOnce(
             () -> {
-              gyro.zeroYaw();
+              drive.updateHeading();
             },
             gyro),
-        new OnePieceAuto(wrist, feeder, shooter),
-        new ParallelCommandGroup(
-            Commands.runOnce(
-                () -> {
-                  wrist.setGoal(0);
-                },
-                wrist),
-            new AllIntakesRun(actuator, otbIntake, utbIntake, feeder, false)),
+        // new OnePieceAuto(wrist, feeder, shooter),
+        // Commands.runOnce(
+        //     () -> {
+        //       wrist.setGoal(0);
+        //     },
+        //     wrist),
+        // new InstantCommand(() -> shooter.setSetpoint(0)),
+        // new ParallelCommandGroup(
         Commands.runOnce(
             () -> {
-              drive.setRaw(0, speed / DriveConstants.MAX_LINEAR_SPEED_M_PER_SEC, 0);
+              drive.driveWithDeadband(0, speed, 0);
+            },
+            drive),
+        // new AllIntakesRun(actuator, otbIntake, utbIntake, feeder, false)),
+        new WaitCommand(seconds / 2),
+        Commands.runOnce(
+            () -> {
+              drive.driveWithDeadband(0, speed / 2, 0);
             },
             drive),
         new WaitCommand(seconds),
         Commands.runOnce(
             () -> {
-              drive.setRaw(0.0, 0, 0);
+              drive.driveWithDeadband(0, -speed, 0);
+              // shooter.setTolerance(500);
             },
             drive),
-        new AllIntakesRun(actuator, otbIntake, utbIntake, feeder, true),
-        new PositionToShoot(
-            feeder, shooter, wrist, WristConstants.PODIUM_RAD, ShooterConstants.CLOSE_RPM));
+        new WaitCommand(seconds),
+        // new AllIntakesRun(actuator, otbIntake, utbIntake, feeder, true),
+        Commands.runOnce(
+            () -> {
+              drive.driveWithDeadband(0, 0, 0);
+            },
+            drive)
+        // ,
+        // new WaitUntilCommand(() -> shooter.allAtSetpoint()),
+        // new Shoot(feeder, arm, shooter),
+        // Commands.runOnce(() -> shooter.setTolerance(ShooterConstants.RPM_TOLERANCE), shooter)
+        );
   }
 }
