+++ src/main/java/frc/robot/Commands/AutonomousCommands/DeadReckons/quickAuto.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.autos;
+package frc.robot.Commands.AutonomousCommands.DeadReckons;
 
 import edu.wpi.first.wpilibj.Timer;
 import frc.robot.Subsystems.drive.Drive;
