+++ src/main/java/frc/robot/Commands/AutonomousCommands/PathPlannerCommands/SpeakerShoot.java
@@ -2,15 +2,14 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Commands.AutonomousCommands.First3Pieces;
+package frc.robot.Commands.AutonomousCommands.PathPlannerCommands;
 
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import edu.wpi.first.wpilibj2.command.WaitCommand;
 import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
+import frc.robot.Commands.TeleopCommands.SpeakerScore.PositionToShoot;
+import frc.robot.Commands.TeleopCommands.SpeakerScore.Shoot;
+import frc.robot.Subsystems.arm.Arm;
 import frc.robot.Subsystems.feeder.Feeder;
-import frc.robot.Subsystems.feeder.FeederConstants;
 import frc.robot.Subsystems.shooter.Shooter;
 import frc.robot.Subsystems.shooter.ShooterConstants;
 import frc.robot.Subsystems.wrist.Wrist;
@@ -19,26 +18,15 @@ import frc.robot.Subsystems.wrist.WristConstants;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class OnePieceAuto extends SequentialCommandGroup {
-  /** Creates a new OnePieceAuto. */
-  public OnePieceAuto(Wrist wrist, Feeder feeder, Shooter shooter) {
+public class SpeakerShoot extends SequentialCommandGroup {
+  /** Creates a new SpeakerShoot. */
+  public SpeakerShoot(Feeder feeder, Shooter shooter, Wrist wrist, Arm arm) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
-        new WaitCommand(SmartDashboard.getNumber("delay", 0)),
-        Commands.runOnce(
-            () -> {
-              wrist.setGoal(WristConstants.SUBWOOFER_RAD);
-              shooter.setSetpoint(ShooterConstants.CLOSE_RPM);
-            },
-            shooter,
-            wrist),
+        new PositionToShoot(
+            feeder, shooter, wrist, WristConstants.SUBWOOFER_RAD, ShooterConstants.CLOSE_RPM),
         new WaitUntilCommand(() -> wrist.atGoal()),
-        new WaitUntilCommand(() -> shooter.topAtSetpoint()),
-        Commands.runOnce(
-            () -> {
-              feeder.setSetpoint(FeederConstants.SPEAKER_RPM);
-            },
-            feeder));
+        new Shoot(feeder, arm, shooter));
   }
 }
