+++ src/main/java/frc/robot/RobotContainer.java
@@ -21,18 +21,18 @@ import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.*;
 import edu.wpi.first.wpilibj2.command.button.*;
-import frc.robot.Commands.AutoCommands.PickUp;
-import frc.robot.Commands.AutoCommands.SpeakerShoot;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.FourPieceBlue;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.FourPieceRed;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.LeaveAuto;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.OnePieceAuto;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.OnePieceLeaveAmpSide;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.OnePieceLeaveCenter;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.OnePieceLeaveCoolSide;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.ThreePieceAutoBlue;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.ThreePieceAutoRed;
-import frc.robot.Commands.AutonomousCommands.First3Pieces.TwoPieceReturnSub;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.FourPieceBlue;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.FourPieceRed;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.LeaveAuto;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.OnePieceAuto;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.OnePieceLeaveAmpSide;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.OnePieceLeaveCenter;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.OnePieceLeaveCoolSide;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.ThreePieceAutoBlue;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.ThreePieceAutoRed;
+import frc.robot.Commands.AutonomousCommands.DeadReckons.First3Pieces.TwoPieceReturnSub;
+import frc.robot.Commands.AutonomousCommands.PathPlannerCommands.PickUp;
+import frc.robot.Commands.AutonomousCommands.PathPlannerCommands.SpeakerShoot;
 import frc.robot.Commands.TeleopCommands.AmpScore.Backside.*;
 import frc.robot.Commands.TeleopCommands.AmpScore.Frontside.*;
 import frc.robot.Commands.TeleopCommands.DefaultDriveCommand;
